requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
async-generator==1.10
atari-py==0.2.6
attrs==20.3.0
bleach==3.2.1
box2d-py==2.3.8
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.1.2
defusedxml==0.6.0
dm-env==1.3
dm-tree==0.1.5
easydict==1.9
entrypoints==0.3
flake8==3.8.4
Flask==1.0.2
future==0.18.2
glfw==2.0.0
gtimer==1.0.0b5
gym==0.18.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.10.2
Jinja2==2.11.2
jsonpatch==1.28
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
labmaze==1.0.3
lockfile==0.12.2
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mujoco-py==1.50.1.68
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.16.0
opencv-python==4.4.0.46
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
pexpect==4.2.1
pickleshare==0.7.4
Pillow==7.2.0
plotly==4.0.0
prometheus-client==0.9.0
prompt-toolkit==1.0.15
protobuf==3.14.0
ptyprocess==0.5.2
pycodestyle==2.6.0
pycparser==2.19
pyflakes==2.2.0
pyglet==1.5.0
Pygments==2.7.4
PyOpenGL==3.1.5
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.6.8
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.1
retrying==1.3.3
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.15.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tld==0.10
torch==1.7.1
torchfile==0.1.0
torchvision==0.8.2
tornado==6.1
tqdm==4.54.1
traitlets==4.3.2
typing-extensions==3.7.4.3
urllib3==1.26.2
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.4.0
